+++ CHANGELOG.md
@@ -10,6 +10,7 @@ v1.11: In progress
 - I have moved the code for the delete element sub tool into the shared section for Form Builder tools.
 - Added the delete sub tool to the Password, Email and Textarea elements. [Feature]
 - Added titles tool to the Form Builder, the title and subtitle can now be set. [Feature]
+- Added button tool to the Form Builder, allows the labels for the submit and reset buttons to be defined. [Feature]
 - Additional tests added to the test suite. [Tests]
 - Minor fixes and improvements.